+++ android/src/main/java/com/uxcam/flutteruxcam/UXOcclusionType.java
@@ -1,0 +1,5 @@
+package com.uxcam.flutteruxcam;
+
+public enum UXOcclusionType {
+    none, occludeAllTextFields, overlay, blur, unknown
+}