+++ ios/Classes/FlutterUxcamPlugin.m
@@ -3,6 +3,16 @@
 
 @import UXCam;
 
+static const NSString *FlutterAppKey = @"userAppKey";
+static const NSString *FlutterEnableMultiSessionRecord = @"enableMultiSessionRecord";
+static const NSString *FlutterEnableScreenNameTagging = @"enableAutomaticScreenNameTagging";
+static const NSString *FlutterEnableCrashHandling = @"enableCrashHandling";
+static const NSString *FlutterEnableNetworkLogs = @"enableNetworkLogging";
+static const NSString *FlutterEnableAdvancedGesture = @"enableAdvancedGestureRecognition";
+static const NSString *FlutterOcclusion = @"occlusion";
+static const NSString *FlutterOccludeScreens = @"screens";
+static const NSString *FlutterExcludeScreens = @"excludeMentionedScreens";
+
 @implementation FlutterUxcamPlugin
 
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar
@@ -13,8 +23,7 @@
 	
     FlutterUxcamPlugin* instance = [[FlutterUxcamPlugin alloc] init];
     [registrar addMethodCallDelegate:instance channel:channel];
-	
-	[UXCam pluginType:@"flutter" version:@"2.0.0"];
+	[UXCam pluginType:@"flutter" version:@"2.1.1"];
 }
 
 // The handler method - this is the entry point from the Dart code
@@ -51,12 +60,126 @@
 	result([@"iOS " stringByAppendingString:UIDevice.currentDevice.systemVersion]);
 }
 
+- (void)startWithConfiguration:(FlutterMethodCall*)call result:(FlutterResult)result
+{
+    NSDictionary *configDict = call.arguments[@"config"];
+    NSString *appKey = configDict[FlutterAppKey];
+    UXCamConfiguration *config = [[UXCamConfiguration alloc] initWithAppKey:appKey];
+    [self updateConfiguration:config withDict:configDict];
+    
+    [UXCam startWithConfiguration:config completionBlock:^(BOOL started) {
+        result(@(started));
+    }];
+}
+
+- (void)applyOcclusion:(FlutterMethodCall*)call result:(FlutterResult)result
+{
+    NSDictionary *occlusionJson = call.arguments[@"occlusion"];
+    if (occlusionJson && ![occlusionJson isKindOfClass:NSNull.class]) {
+        id <UXCamOcclusionSetting> setting = [UXCamOcclusion getSettingFromJson:occlusionJson];
+        if (setting)
+        {
+            [UXCam applyOcclusion:setting];
+        }
+    }
+}
+
+- (void)removeOcclusion:(FlutterMethodCall*)call result:(FlutterResult)result
+{
+    NSDictionary *occlusionJson = call.arguments[@"occlusion"];
+    if (occlusionJson && ![occlusionJson isKindOfClass:NSNull.class]) {
+        id <UXCamOcclusionSetting> setting = [UXCamOcclusion getSettingFromJson:occlusionJson];
+        if (setting)
+        {
+            [UXCam removeOcclusionOfType:setting.type];
+        }
+        else
+        {
+            [UXCam removeOcclusion];
+        }
+    }
+}
+
+- (void)updateConfiguration:(FlutterMethodCall*)call result:(FlutterResult)result
+{
+    NSDictionary *configDict = call.arguments[@"config"];
+    UXCamConfiguration *config = UXCam.configuration;
+    if (!config)
+    {
+        NSLog(@"Please call startWithConfiguration: before updating configuration");
+        return;
+    }
+    
+    [self updateConfiguration:config withDict:configDict];
+}
+
+- (void)updateConfiguration:(UXCamConfiguration *)config withDict:(NSDictionary *)configDict
+{
+    NSNumber *enableMultiSessionRecord = configDict[FlutterEnableMultiSessionRecord];
+    if (enableMultiSessionRecord && ![enableMultiSessionRecord isKindOfClass:NSNull.class]) {
+        config.enableMultiSessionRecord = [enableMultiSessionRecord boolValue];
+    }
+    
+    NSNumber *enableCrashHandling = configDict[FlutterEnableCrashHandling];
+    if (enableCrashHandling && ![enableCrashHandling isKindOfClass:NSNull.class]) {
+        config.enableCrashHandling = [enableCrashHandling boolValue];
+    }
+    NSNumber *enableAutomaticScreenNameTagging = configDict[FlutterEnableScreenNameTagging];
+    if (enableAutomaticScreenNameTagging && ![enableAutomaticScreenNameTagging isKindOfClass:NSNull.class]) {
+        config.enableAutomaticScreenNameTagging = [enableAutomaticScreenNameTagging boolValue];
+    }
+    NSNumber *enableNetworkLogging = configDict[FlutterEnableNetworkLogs];
+    if (enableNetworkLogging && ![enableNetworkLogging isKindOfClass:NSNull.class]) {
+        config.enableNetworkLogging = [enableNetworkLogging boolValue];
+    }
+    NSNumber *enableAdvancedGestureRecognition = configDict[FlutterEnableAdvancedGesture];
+    if (enableAdvancedGestureRecognition && ![enableAdvancedGestureRecognition isKindOfClass:NSNull.class]) {
+        config.enableAdvancedGestureRecognition = [enableAdvancedGestureRecognition boolValue];
+    }
+
+    NSArray *occlusionList = configDict[FlutterOcclusion];
+    if (occlusionList && ![occlusionList isKindOfClass:NSNull.class]) {
+        UXCamOcclusion *occlusion = [[UXCamOcclusion alloc] init];
+        for (NSDictionary *occlusionJson in occlusionList) {
+            id <UXCamOcclusionSetting> setting = [UXCamOcclusion getSettingFromJson:occlusionJson];
+            if (setting)
+            {
+                NSArray *screens = occlusionJson[FlutterOccludeScreens];
+                BOOL excludeMentionedScreens = [occlusionJson[FlutterExcludeScreens] boolValue];
+                [occlusion applySettings:@[setting] screens:screens excludeMentionedScreens: excludeMentionedScreens];
+            }
+        }
+        config.occlusion = occlusion;
+    }
+}
+
+- (void)configurationForUXCam:(FlutterMethodCall*)call result:(FlutterResult)result
+{
+    UXCamConfiguration *configuration = UXCam.configuration;
+    if (!configuration)
+    {
+        NSLog(@"Please call startWithConfiguration: before fetching configuration");
+        return result(nil);
+    }
+    
+    NSDictionary *configDict = @{
+        FlutterAppKey: configuration.userAppKey,
+        FlutterEnableMultiSessionRecord: @(configuration.enableMultiSessionRecord),
+        FlutterEnableScreenNameTagging: @(configuration.enableAutomaticScreenNameTagging),
+        FlutterEnableCrashHandling: @(configuration.enableCrashHandling),
+        FlutterEnableAdvancedGesture: @(configuration.enableAdvancedGestureRecognition),
+        FlutterEnableNetworkLogs: @(configuration.enableNetworkLogging)
+    };
+    result(configDict);
+}
+
 - (void)startWithKey:(FlutterMethodCall*)call result:(FlutterResult)result
 {
-	NSString* apiKey = call.arguments[@"key"];
-	[UXCam startWithKey:apiKey completionBlock:^(BOOL started) {
-		result(@(started));
-	}];
+	NSString* appKey = call.arguments[@"key"];
+    UXCamConfiguration *config = [[UXCamConfiguration alloc] initWithAppKey:appKey];
+    [UXCam startWithConfiguration:config completionBlock:^(BOOL started) {
+        result(@(started));
+    }];
 }
 
 - (void)startNewSession:(FlutterMethodCall*)call result:(FlutterResult)result
@@ -67,8 +190,9 @@
 
 - (void)stopSessionAndUploadData:(FlutterMethodCall*)call result:(FlutterResult)result
 {
-	[UXCam stopSessionAndUploadData];
-	result(nil);
+	[UXCam stopSessionAndUploadData:^{
+           result(nil);
+    }];
 }
 
 - (void)allowShortBreakForAnotherApp:(FlutterMethodCall*)call result:(FlutterResult)result