+++ lib/flutter_uxcam.dart
@@ -1,6 +1,63 @@
 import 'dart:async';
 import 'dart:io';
 import 'package:flutter/services.dart';
+import 'package:flutter_uxcam/flutter_occlusion.dart';
+
+class FlutterUxConfigKeys {
+  static const userAppKey = "userAppKey";
+  static const enableMultiSessionRecord = "enableMultiSessionRecord";
+  static const enableCrashHandling = "enableCrashHandling";
+  static const enableAutomaticScreenNameTagging = "enableAutomaticScreenNameTagging";
+  static const enableNetworkLogging = "enableNetworkLogging";
+  static const enableAdvancedGestureRecognition = "enableAdvancedGestureRecognition";
+  static const occlusion = "occlusion";
+}
+
+
+class FlutterUxConfig {
+  String userAppKey;
+
+  // If value is not specified for below variables, default values will be applied as per sdk
+  bool? enableMultiSessionRecord;
+  bool? enableCrashHandling;
+  bool? enableAutomaticScreenNameTagging;
+  bool? enableNetworkLogging;
+  bool? enableAdvancedGestureRecognition;
+  List<FlutterUXOcclusion>? occlusions;
+
+  FlutterUxConfig({
+    required this.userAppKey,
+    this.enableMultiSessionRecord,
+    this.enableCrashHandling,
+    this.enableAutomaticScreenNameTagging,
+    this.enableNetworkLogging,
+    this.enableAdvancedGestureRecognition,
+    this.occlusions
+  });
+
+  factory FlutterUxConfig.fromJson(Map<String, dynamic> json) {
+      var userAppKey = json[FlutterUxConfigKeys.userAppKey];
+      var config = FlutterUxConfig(userAppKey: userAppKey);
+      config.enableMultiSessionRecord = json[FlutterUxConfigKeys.enableMultiSessionRecord];
+      config.enableCrashHandling = json[FlutterUxConfigKeys.enableCrashHandling];
+      config.enableAutomaticScreenNameTagging = json[FlutterUxConfigKeys.enableAutomaticScreenNameTagging];
+      config.enableNetworkLogging = json[FlutterUxConfigKeys.enableNetworkLogging];
+      config.enableAdvancedGestureRecognition = json[FlutterUxConfigKeys.enableAdvancedGestureRecognition];
+      return config;
+  }
+
+  Map<String, dynamic> toJson() {
+    return {
+      FlutterUxConfigKeys.userAppKey: userAppKey,
+      FlutterUxConfigKeys.enableMultiSessionRecord: enableMultiSessionRecord,
+      FlutterUxConfigKeys.enableCrashHandling: enableCrashHandling,
+      FlutterUxConfigKeys.enableAutomaticScreenNameTagging: enableAutomaticScreenNameTagging,
+      FlutterUxConfigKeys.enableNetworkLogging: enableNetworkLogging,
+      FlutterUxConfigKeys.enableAdvancedGestureRecognition: enableAdvancedGestureRecognition,
+      FlutterUxConfigKeys.occlusion: occlusions?.map((occlusion) => occlusion.toJson()).toList()
+    };
+  }
+}
 
 class FlutterUxcam {
   static const MethodChannel _channel = const MethodChannel('flutter_uxcam');
@@ -9,6 +66,23 @@ class FlutterUxcam {
     final String? version =
         await _channel.invokeMethod<String>('getPlatformVersion');
     return version!;
+  }
+
+  static Future<bool> startWithConfiguration(FlutterUxConfig config) async {
+    final bool? status =
+    await _channel.invokeMethod<bool>('startWithConfiguration', {"config": config.toJson()});
+    return status!;
+  }
+
+  static Future<FlutterUxConfig> configurationForUXCam() async {
+    final Map<String, dynamic>? json = await _channel.invokeMapMethod('configurationForUXCam');
+    return FlutterUxConfig.fromJson(json!);
+  }
+
+  static Future<bool> updateConfiguration(FlutterUxConfig config) async {
+    final bool? status =
+    await _channel.invokeMethod<bool>('updateConfiguration', {"config": config.toJson()});
+    return status!;
   }
 
   static Future<bool> startWithKey(String key) async {
@@ -48,7 +122,7 @@ class FlutterUxcam {
     await _channel.invokeMethod('tagScreenName', {"key": screenName});
   }
 
-  static Future<void> setUserIdentity(String userIdentity) async {
+  static Future<void> setUserIdentity(String? userIdentity) async {
     await _channel.invokeMethod('setUserIdentity', {"key": userIdentity});
   }
 
@@ -176,6 +250,7 @@ class FlutterUxcam {
     await _channel.invokeMethod('uploadPendingSession');
   }
 
+  @Deprecated("Please use stopSessionAndUploadData() instead")
   static Future<void> stopApplicationAndUploadData() async {
     await _channel.invokeMethod('stopApplicationAndUploadData');
   }
@@ -213,5 +288,21 @@ class FlutterUxcam {
       [Map<String, dynamic>? properties]) async {
     await _channel.invokeMethod(
         'reportBugEvent', {"eventName": eventName, "properties": properties});
+  }
+
+  static Future<bool> applyOcclusion(FlutterUXOcclusion occlusion) async {
+    final bool? status = await _channel.invokeMethod<bool>(
+        'applyOcclusion',
+        { "occlusion": occlusion.toJson()}
+    );
+    return status!;
+  }
+
+  static Future<bool> removeOcclusion(FlutterUXOcclusion occlusion) async {
+    final bool? status = await _channel.invokeMethod<bool>(
+        'removeOcclusion',
+        { "occlusion": occlusion.toJson() }
+    );
+    return status!;
   }
 }